Crate commodore

Call rank and take command of Slack with rust at your helm

Structs

Attachment

AttachmentBuilder

Command
A struct representation of a Slack Command and the context from which it was triggered

Field

MatchCommand
A direct command matcher

MatchSubCommand
A matcher that assumes any text starting with the provided string is a subcommand. i.e. /cmd help

MatchText
A regex pattern matcher for command text. Regex captures will be provided to the matched Handler

Mux
A command de-multiplexor

Response
A payload to reply to commands with

ResponseBuilder

Traits

Handler
Command handling interface Implementation for Fn

Matcher
Command matching interface

Responder
Deferred response interface

Type Definitions

Captures
Results for regex matchers that collect captures