Struct commodore::Mux

pub struct Mux { /* fields omitted */ }

A command de-multiplexor

Methods

impl Mux

fn new() -> Mux

fn command<C, T, H>(&mut self, cmd: C, token: T, handler: H) where C: Into<String>, T: Into<String>, H: Handler + 'static

Install routing for a Slack command, secret token, and target Handler

fn matching<M, H>(&mut self, matcher: M, handler: H) where M: Matcher + 'static, H: Handler + 'static

Install routing for a Slack command matcher and target Handler

fn route(&mut self, route: Route)

Install a command routing

fn handler<'a>(&self,
cmd: &'a Command)
-> Option<(Option<Captures<'a>>, &Box<Handler>)>

Attempts to return the first match result for a target Handler